backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
playwright==1.40.0
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
alembic==1.13.0
redis==5.0.1
celery==5.3.4
pandas==2.1.4
numpy==1.25.2
aiofiles==23.2.1
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1 